tools/cross_compile_android/requirements.txt
gitpython
Jinja2
tqdm
